Use SoftReference instead of WeakReference for caching
My original use case for these inline caches was to go in front of a hash-based
cache, and so a WeakReference made sense: the referent will stop being useful
at the same time it stops being reachable. But actually they wind up being
used more generally, including to cache things nobody else has a reference to,
such as a Set of hash-cached items, or Type objects, which I think are not
interned, and so we save some time by asking the GC to keep them around.

PiperOrigin-RevId: 331587674