Suppress @Var check for classes with @Generated annotation.
This disables @Var checking of auto-generated classes. Enforcing the check for
generated code isn't very useful, because this code is not written and rarely
read by humans.

The practical implication is that projects that rely on frameworks that generate
code will no longer see warnings about the auto-generated code, which also means
that the check can be enabled at error level severity for those projects, since
the only errors remaining are in the human-written code which should be fixed.

PiperOrigin-RevId: 347580651