Introduce new helpers for testonly file objects and file managers
to improve test performance and eventually replace
`ErrorProneInMemoryFileManager`.

Currently we create a fresh `ErrorProneInMemoryFileManager` for each test
method, which re-does expensive IO to read the platform and user classpaths. We
don't expect either of those to change, so this change introduces a single
persistent `JavacFileManager` cached in a `ThreadLocal` that can be re-used by
multiple tests. It also simplifies test input
creation by introducing utilities for managing in-memory file
objects for test inputs, which are no longer tied to the
`ErrorProneInMemoryFileManager` abstraction.

PiperOrigin-RevId: 348516038