Implement BanSerializableRead check (see buganizer/162956872).
This CL tries to address bug 162956872 by looking for places where arbitrary java entities are deserialized. It does this by banning calls to:

- `java.io.ObjectInputStream.readObject`
- `java.io.ObjectInputStream.defaultReadObject`
- `java.io.ObjectInputStream.readClassDescriptor`
- `java.io.ObjectInputStream.resolveClass`
- `java.io.ObjectInputStream.resolveObject`

Many classes define a `readObject()` method that defines how a class *would* deserialise, if it *were* deserialised. To allow this, we exempt these checks if the containing function is a `readObject()` method, then block actual calls to the `readObject()` method itself.

It also introduces several suppression directives:

- `SuppressBanSerializableCompletedSecurityReview`: exempted because ISE reviewed this
- `SuppressBanSerializableForLegacyCode`: exempted for legacy code
- `DangerouslySuppressBanSerializable`: proposed as something for external google users to use ... I am not sure if they'd have much context on 'security review', and I want to make it clear that it's dangerous

A refactoring is also implemented, which suggests adding `@DangerouslySuppressBanSerializable`.

Open questions for reviewers:

- What kind of API are we exposing to external Google users? I am not sure that having our API generate legacy or security review annotations as a 'fix' makes much sense. Maybe there is precedent on this?

PiperOrigin-RevId: 349609882